本番環境でコンポーネント置き場が見えないように
